+++ UniversiteDomain/Entities/Notes.cs
@@ -2,7 +2,7 @@
 {
     public class Notes
     {
-        public double Valeur { get; set; }
+        public float Valeur { get; set; }
         public long EtudiantId { get; set; }
         public Etudiant Etudiant { get; set; }
